+++ src/config/Blocks/listingVariations.js
@@ -185,17 +185,17 @@ const italiaListingVariations = [
       return schema;
     },
   },
-//  {
-//    id: 'amministrazioneTrasparenteTablesTemplate',
-//    isDefault: false,
-//    title: 'Tabelle Amministrazione Trasparente',
-//    template: AmministrazioneTrasparenteTablesTemplate,
-//    skeleton: AmministrazioneTrasparenteTablesTemplateSkeleton,
-//    schemaEnhancer: ({ schema, formData, intl }) => {
-//      /*let pos = */ addDefaultOptions(schema, formData, intl);
-//      return schema;
-//    },
-  },
+  //  {
+  //    id: 'amministrazioneTrasparenteTablesTemplate',
+  //    isDefault: false,
+  //    title: 'Tabelle Amministrazione Trasparente',
+  //    template: AmministrazioneTrasparenteTablesTemplate,
+  //    skeleton: AmministrazioneTrasparenteTablesTemplateSkeleton,
+  //    schemaEnhancer: ({ schema, formData, intl }) => {
+  //      /*let pos = */ addDefaultOptions(schema, formData, intl);
+  //      return schema;
+  //    },
+  // },
 
   // ****** Example: ******
   // { id:template_id,